Add initial applications and updated methods
I have added applications of the data in section 4, including data visualizations and some statistical analysis. I will need to perform additional analysis regarding TSP threshholds, but am uncertain at this time how to proceed with that analysis. I also made minor modifications to the methods section so that it will knit with the plots derived directly in the code.